requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beanie==1.27.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.6.0
orjson==3.10.7
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.1
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
toml==0.10.2
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
